// angular_fe/src/app/interfaces/dashboard/categoryStatsDTO.interface.ts
/**
 * @category Interfaces
 * 
 * @description Represents statistical data for a product category.
 * 
 * @interface CategoryStatsDTO
 * @property {number} id - The unique identifier of the category.
 * @property {string} name - The name of the category.
 * @property {number} revenue - Total revenue generated by this category.
 * @property {number} productCount - Total number of products in this category.
 * @property {number} orderCount - Total number of orders containing products from this category.
 * @property {number} percentageOfTotal - Percentage of total revenue this category represents.
 * 
 * @author Popa Madalina Mariana
 * @version 0.0.0
 */
export interface CategoryStatsDTO {
  id: number;
  name: string;
  revenue: number;
  productCount: number;
  orderCount: number;
  percentageOfTotal: number;
}